requirements.txt
python-dotenv>=1.0.0
tabulate==0.9.0